Split up SK*StoreProduct implementations in separate files
