+++ Purchases/Purchasing/StoreKitAbstractions/StoreProduct.swift
@@ -211,160 +211,3 @@ extension StoreProduct {
     }
 
 }
-
-// MARK: - Subclasses
-
-@available(iOS 15.0, tvOS 15.0, watchOS 8.0, macOS 12.0, *)
-internal struct SK2StoreProduct: StoreProductType {
-
-    init(sk2Product: SK2Product) {
-        self._underlyingSK2Product = sk2Product
-    }
-
-    // We can't directly store instances of StoreKit.Product, since that causes
-    // linking issues in iOS < 15, even with @available checks correctly in place.
-    // So instead, we store the underlying product as Any and wrap it with casting.
-    // https://openradar.appspot.com/radar?id=4970535809187840
-    private let _underlyingSK2Product: Any
-    var underlyingSK2Product: SK2Product {
-        // swiftlint:disable:next force_cast
-        _underlyingSK2Product as! SK2Product
-    }
-
-    var localizedDescription: String { underlyingSK2Product.description }
-
-    var price: Decimal { underlyingSK2Product.price }
-
-    var localizedPriceString: String { underlyingSK2Product.displayPrice }
-
-    var productIdentifier: String { underlyingSK2Product.id }
-
-    var isFamilyShareable: Bool { underlyingSK2Product.isFamilyShareable }
-
-    var localizedTitle: String { underlyingSK2Product.displayName }
-
-    var priceFormatter: NumberFormatter? {
-        // note: if we ever need more information from the jsonRepresentation object, we
-        // should use Codable or another decoding method to clean up this code.
-        guard let attributes = jsonDict["attributes"] as? [String: Any],
-              let offers = attributes["offers"] as? [[String: Any]],
-              let currencyCode: String = offers.first?["currencyCode"] as? String else {
-            Logger.appleError("Can't initialize priceFormatter for SK2 product! Could not find the currency code")
-            return nil
-        }
-        let formatter = NumberFormatter()
-        formatter.numberStyle = .currency
-        formatter.currencyCode = currencyCode
-        formatter.locale = .autoupdatingCurrent
-        return formatter
-    }
-
-    var subscriptionGroupIdentifier: String? {
-        underlyingSK2Product.subscription?.subscriptionGroupID
-    }
-
-    private var jsonDict: [String: Any] {
-        let decoded = try? JSONSerialization.jsonObject(with: self.underlyingSK2Product.jsonRepresentation, options: [])
-        return decoded as? [String: Any] ?? [:]
-    }
-
-    var subscriptionPeriod: SubscriptionPeriod? {
-        guard let skSubscriptionPeriod = underlyingSK2Product.subscription?.subscriptionPeriod else {
-            return nil
-        }
-        return SubscriptionPeriod.from(sk2SubscriptionPeriod: skSubscriptionPeriod)
-    }
-
-    var introductoryPrice: PromotionalOffer? {
-        self.underlyingSK2Product.subscription?.introductoryOffer
-            .map(PromotionalOffer.init)
-    }
-
-    var discounts: [PromotionalOffer] {
-        (self.underlyingSK2Product.subscription?.promotionalOffers ?? [])
-            .compactMap(PromotionalOffer.init)
-    }
-
-}
-
-internal struct SK1StoreProduct: StoreProductType {
-
-    init(sk1Product: SK1Product) {
-        self.underlyingSK1Product = sk1Product
-    }
-
-    let underlyingSK1Product: SK1Product
-
-    var localizedDescription: String { return underlyingSK1Product.localizedDescription }
-
-    var price: Decimal { return underlyingSK1Product.price as Decimal }
-
-    var localizedPriceString: String {
-        return priceFormatter?.string(from: underlyingSK1Product.price) ?? ""
-    }
-
-    var productIdentifier: String { return underlyingSK1Product.productIdentifier }
-
-    @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 8.0, *)
-    var isFamilyShareable: Bool { underlyingSK1Product.isFamilyShareable }
-
-    var localizedTitle: String { underlyingSK1Product.localizedTitle }
-
-    @available(iOS 12.0, macCatalyst 13.0, tvOS 12.0, macOS 10.14, watchOS 6.2, *)
-    var subscriptionGroupIdentifier: String? { underlyingSK1Product.subscriptionGroupIdentifier }
-
-    var priceFormatter: NumberFormatter? {
-        let formatter = NumberFormatter()
-        formatter.numberStyle = .currency
-        formatter.locale = underlyingSK1Product.priceLocale
-        return formatter
-    }
-
-    @available(iOS 11.2, macOS 10.13.2, tvOS 11.2, watchOS 6.2, *)
-    var subscriptionPeriod: SubscriptionPeriod? {
-        guard let skSubscriptionPeriod = underlyingSK1Product.subscriptionPeriod else {
-            return nil
-        }
-        return SubscriptionPeriod.from(sk1SubscriptionPeriod: skSubscriptionPeriod)
-    }
-
-    @available(iOS 12.2, macOS 10.14.4, tvOS 12.2, watchOS 6.2, *)
-    var introductoryPrice: PromotionalOffer? {
-        return self.underlyingSK1Product.introductoryPrice
-            .map(PromotionalOffer.init)
-    }
-
-    @available(iOS 12.2, macOS 10.14.4, tvOS 12.2, watchOS 6.2, *)
-    var discounts: [PromotionalOffer] {
-        return self.underlyingSK1Product.discounts
-            .map(PromotionalOffer.init)
-    }
-
-}
-
-// MARK: - Extensions
-
-extension SK1StoreProduct: Hashable {
-
-    static func == (lhs: SK1StoreProduct, rhs: SK1StoreProduct) -> Bool {
-        return lhs.underlyingSK1Product == rhs.underlyingSK1Product
-    }
-
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(self.underlyingSK1Product)
-    }
-
-}
-
-@available(iOS 15.0, tvOS 15.0, watchOS 8.0, macOS 12.0, *)
-extension SK2StoreProduct: Hashable {
-
-    static func == (lhs: SK2StoreProduct, rhs: SK2StoreProduct) -> Bool {
-        return lhs.underlyingSK2Product == rhs.underlyingSK2Product
-    }
-
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(self.underlyingSK2Product)
-    }
-
-}
